BUG: Reidentify doesn't enable after seed ident.
The previous change disabled Reidentify until after seed is identified. The check to enable the Reidentify button was running before the seed field was populated, meaning that the button would only be enabled after the second completion of Monitor Blinks. This fix adjusts the timing so it happens after the first completion of Monitor Blinks.